# swagger.yml
openapi: 3.1.0
info:
  title: CatMapper API
  version: 1.1.0
  description: |
    Public and authenticated endpoints for CatMapper (SocioMap and ArchaMap).

    Canonical REST endpoints are now exposed under `/api`. Existing unprefixed
    routes remain available as deprecated compatibility aliases during the
    migration.

    Notes on auth:
    - Most read endpoints do not require authentication.
    - Write/task endpoints support either:
      1) `Authorization: Bearer <token>`
      2) `X-API-Key: <key>` (or `Authorization: ApiKey <key>`)
      3) Backward-compatible body credentials in `cred` (`{ userid, key }`).
servers:
  - url: https://api.catmapper.org/api
    description: Production canonical REST API
  - url: http://localhost:5000/api
    description: Local development canonical REST API
...
paths:
...
  /translation-tasks:
    post:
      tags: [Translate]
      summary: Start an asynchronous translation task
      description: 'Canonical REST endpoint. Deprecated compatibility alias: `/translate/start`.'
      responses:
        '200':
          description: Translation task state
        '500':
          $ref: '#/components/responses/InternalError'
...
  /search:
    get:
      tags: [Search]
      summary: Search categories/entities
      description: |
        Searches a CatMapper database for matching categories/entities.
        This is the canonical REST path when used with the `/api` server base;
        the unprefixed `/search` route remains as a deprecated compatibility alias.

        Multi-context filtering:
        - `contexts` can be passed as comma-separated values and/or repeated params.
        - `contextMode=all` requires ties to every context CMID.
        - `contextMode=any` requires at least one context CMID.
      parameters:
        - $ref: '#/components/parameters/databaseQuery'
        - name: term
          in: query
          required: false
          schema:
            type: string
          description: Search term.
        - name: property
          in: query
          required: false
          schema:
            type: string
            enum: [Name, CMID, Key]
          description: Property to search by.
        - name: domain
          in: query
          required: false
          schema:
            type: string
            default: CATEGORY
          description: Domain/subdomain label to constrain matches.
        - name: dataset
          in: query
          required: false
          schema:
            type: string
          description: Optional dataset CMID filter.
        - name: yearStart
          in: query
          required: false
          schema:
            oneOf:
              - type: integer
              - type: string
        - name: yearEnd
          in: query
          required: false
          schema:
            oneOf:
              - type: integer
              - type: string
        - name: country
          in: query
          required: false
          schema:
            type: string
          description: Context by ADM0 CMID.
        - name: context
          in: query
          required: false
          schema:
            type: string
          description: Single context CMID.
        - name: contexts
          in: query
          required: false
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
          description: Comma-separated or repeated context CMIDs.
        - name: contextMode
          in: query
          required: false
          schema:
            type: string
            enum: [all, any]
            default: all
          description: How multiple contexts are combined.
        - name: limit
          in: query
          required: false
          schema:
            oneOf:
              - type: integer
              - type: string
            default: 10000
        - name: query
          in: query
          required: false
          schema:
            type: string
            enum: ['true', 'false']
          description: Return Cypher query metadata instead of result payload details.
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        '500':
          $ref: '#/components/responses/InternalError'
  /translate:
    post:
      tags: [Translate]
      summary: Synchronous translate
      description: Performs translation and returns rows immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
      responses:
        '200':
          description: Translation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /translate/start:
    post:
      tags: [Translate]
      summary: Start async translate task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TranslateRequest'
                - type: object
                  properties:
                    batchSize:
                      type: integer
                      minimum: 1
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateTask'
        '500':
          $ref: '#/components/responses/InternalError'
  /translate/status:
    post:
      tags: [Translate]
      summary: Get async translate status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId]
              properties:
                taskId:
                  type: string
      responses:
        '200':
          description: Task state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateTask'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /translate/cancel:
    post:
      tags: [Translate]
      summary: Cancel async translate task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId]
              properties:
                taskId:
                  type: string
      responses:
        '200':
          description: Task state after cancel request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateTask'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /nlp/parse-log:
    post:
      tags: [Translate]
      summary: Save NLP parse request log entry
      description: Saves a JSON log entry under server log storage for NLP parse diagnostics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Log entry written
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  path:
                    type: string
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /dataset:
    get:
      tags: [Explore]
      summary: Dataset detail/query (GET)
      parameters:
        - $ref: '#/components/parameters/databaseQuery'
        - name: cmid
          in: query
          required: true
          schema:
            type: string
        - name: domain
          in: query
          required: false
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: children
          in: query
          required: false
          schema:
            oneOf:
              - type: boolean
              - type: string
      responses:
        '200':
          description: Dataset payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    databaseQuery:
      name: database
      in: query
      required: true
      schema:
        type: string
      description: Database identifier (for example `SocioMap` or `ArchaMap`).
...
  responses:
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_details:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                nullable: true
              field:
                type: string
              code:
                type: string
              message:
                type: string
            additionalProperties: true
      additionalProperties: true
...
    SearchResult:
      type: object
      properties:
        CMID:
          type: string
        CMName:
          type: string
        country:
          type: array
          items:
            type: string
        domain:
          type: array
          items:
            type: string
        matching:
          type: string
        matchingDistance:
          oneOf:
            - type: integer
            - type: number
      additionalProperties: true
    TranslateRequest:
      type: object
      properties:
        database:
          type: string
        property:
          type: string
        domain:
          type: string
        key:
          type: string
        term:
          type: string
        country:
          type: string
        context:
          type: string
        dataset:
          type: string
        yearStart:
          oneOf:
            - type: integer
            - type: string
        yearEnd:
          oneOf:
            - type: integer
            - type: string
        query:
          oneOf:
            - type: boolean
            - type: string
        table:
          type: array
          items:
            type: object
            additionalProperties: true
        countsamename:
          oneOf:
            - type: boolean
            - type: string
        uniqueRows:
          oneOf:
            - type: boolean
            - type: string
      additionalProperties: true
    TranslateResponse:
      type: object
      properties:
        file:
          type: array
          items:
            type: object
            additionalProperties: true
        order:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      additionalProperties: true
    TranslateTask:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
          enum: [processing, completed, failed, canceled, queued]
        stage:
          type: string
        percent:
          type: integer
        message:
          type: string
        error:
          type: string
        file:
          type: array
          items:
            type: object
            additionalProperties: true
        order:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        processedRows:
          type: integer
        totalRows:
          type: integer
        elapsedSeconds:
          type: number
      additionalProperties: true
...